requirements.txt
Flask==3.0.3
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
Werkzeug==3.0.3
